feat: implement UI configuration modal and refactor UI state management
